+++ README.md
@@ -1,7 +1,13 @@
 ATOMIC WORKOUTS
 
+# Branch Styling
+
+- make muscles like hashtags visually
+- check all pages and find things to design
+- Branch: Timer Landing
 
 # Offline Mode Setup
+
 fill offline database:
 mongoimport --db 3 --collection workouts --file workouts.json --jsonArray
 mongoimport --db 3 --collection exercises --file exercises.json --jsonArray
